requirements.txt
ACSConv
matplotlib
ffmpeg
nibabel
nipype
numpy
opencv-python
pandas
performer-pytorch
scikit-image
scikit-learn
scipy
SimpleITK
timm
torchio
torchvision
umap-learn
umap-learn[plot]
pytorch-lightning
pytorch-metric-learning